Fixed navigation, is apparently pretty easy, don't know why we were told not to do it
