--- Windows-App/Navigation/INavigateablePage.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Windows.UI.Xaml.Controls;
-
-namespace Windows_App.Navigation
-{
-    interface INavigateablePage
-    {
-       SplitView ContainerSplitView{ get;}
-    }
-}
