--- Windows-App/Navigation/NavigationElement.xaml
@@ -1,41 +1,0 @@
-﻿<Page
-    x:Class="Windows_App.NavigationElement"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Windows_App"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-    <StackPanel Background="Gray">
-        <Button 
-                        x:Name="NavCollapseButton"
-                        FontFamily="Segoe MDL2 Assets" 
-                        Content="&#xE700;" 
-                        Click="NavCollapseButton_Click"
-                        Width="50"
-                        Height="50"/>
-        <StackPanel x:Name="NavPageBusinessesStackPanel" Orientation="Horizontal" Tapped="NavPageBusinessesStackPanel_Tapped"  >
-            <Button FontFamily="Segoe MDL2 Assets" 
-                            Content="&#xE913;"
-                            Width="50"
-                            Height="50"/>
-            <TextBlock Text="Ondernemingen" FontSize="18" VerticalAlignment="Center" Margin="10,0,0,0"/>
-        </StackPanel>
-        <StackPanel x:Name="NavPagePromotionsStackPanel" Orientation="Horizontal" Tapped="NavPagePromotionsStackPanel_Tapped"   >
-            <Button FontFamily="Segoe MDL2 Assets" 
-                            Content="&#xE728;"
-                            Width="50"
-                            Height="50"/>
-            <TextBlock Text="Promoties" FontSize="18" VerticalAlignment="Center" Margin="13,0,0,0"/>
-        </StackPanel>
-        <StackPanel x:Name="NavPageEventsStackPanel" Orientation="Horizontal" Tapped="NavPageEventsStackPanel_Tapped"   >
-            <Button FontFamily="Segoe MDL2 Assets" 
-                            Content="&#xE787;"
-                            Width="50"
-                            Height="50"/>
-            <TextBlock Text="Evenementen" FontSize="18" VerticalAlignment="Center" Margin="10,0,0,0"/>
-        </StackPanel>
-    </StackPanel>
-</Page>
