+++ Windows-App/View/PromotionsPage.xaml
@@ -9,14 +9,7 @@
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
-    <SplitView x:Name="PageSplitView" DisplayMode="CompactOverlay">
-        <SplitView.Pane>
-            <local1:NavigationElement />
-        </SplitView.Pane>
-        <SplitView.Content>
-            <StackPanel>
-                <TextBlock Text="Promotions"/>
-            </StackPanel>
-        </SplitView.Content>
-    </SplitView>
+    <StackPanel>
+        <TextBlock Text="Promotions"/>
+    </StackPanel>
 </Page>
