+++ Windows-App/Windows-App.csproj
@@ -103,10 +103,6 @@
     <Compile Include="Model\Promotion.cs" />
     <Compile Include="Model\Establishment.cs" />
     <Compile Include="Model\User.cs" />
-    <Compile Include="Navigation\INavigateablePage.cs" />
-    <Compile Include="Navigation\NavigationElement.xaml.cs">
-      <DependentUpon>NavigationElement.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="View\BusinessesPage.xaml.cs">
       <DependentUpon>BusinessesPage.xaml</DependentUpon>
@@ -142,10 +138,6 @@
     <Page Include="MainPage.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
-    </Page>
-    <Page Include="Navigation\NavigationElement.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="View\BusinessesPage.xaml">
       <SubType>Designer</SubType>
